Moved all test classes to ct3PhoneBookTest
